pyproject.toml
[tool.poetry]
name = "deltatorch"
version = "0.0.1"
description = "DeltaTorch allows loading training data from  DeltaLake tables for training Deep Learning models using PyTorch"
authors = ["Michael Shtelma <mshtelma@gmail.com>"]
license = "Apache-2.0"
readme = "README.md"
homepage = "https://github.com/mshtelma/deltatorch/"
keywords = ['delta', 'torch', 'pytorch', 'deltalake']

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
deltalake = ">=0.8"
pillow = ">=9.5"

[tool.poetry.dev-dependencies]
setuptools = ">=67.7.2"
requests = "2.29.0"
torchvision = ">=0.15"
pandas = ">=2.0.0"
pyspark = ">=3.4.0"
pytest = ">=7.3.0"
pytest-describe = ">=2"
portalocker = ">=2"
black= ">=23.3.0"
flake8-bugbear = ">=23"
flake8-pyprojecttoml = ">=0.0.2"
flake8-pyproject = ">=1.2.3"
pep8-naming = ">=0.13"
diffusers = ">=0.16.1"
accelerate = ">=0.18.0"
transformers = ">=4.29.1"
datasets = ">=2.12.0"
pytorch-lightning = ">=2.0.2"
torchmetrics = ">=0.11.4"


[tool.poetry.group.mkdocs]
optional = true

[tool.poetry.group.mkdocs.dependencies]
mkdocstrings-python = "^0.8.3"
mkdocs-gen-files = "^0.4.0"
mkdocs-literate-nav = "^0.6.0"
mkdocs-section-index = "^0.3.5"
markdown-include = "^0.8.1"
mkdocs = "^1.4.2"



[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"